fix: behav_eq and relatedness??
fixed `behav_eq` when run for populations with only two individuals. I don't know why this works. It really makes no sense.
add negative checks for the sqrts in `benefit`.
add multiplier term to benefit to pretend two individuals interacting is actually ten. This is so punishment has the intended effect when related individuals interact.
add test for fitness and payoff.
update test for `behav_eq`.

+++ CoopConflGameStructs.jl
@@ -10,7 +10,7 @@ mutable struct simulation_parameters
     T0::Float64
     #popgen params
     gmax::Int64  # maximum number of generations
-    tmax::Float32  # maximum length of timespan for ODE
+    tmax::Float64  # maximum length of timespan for ODE
     population_size::Int64
     synergy::Float64
     relatedness::Float64
